# src/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "NFT Auction Backend API Documentation",
        "title": "NFT Auction API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.nftauction.com/support",
            "email": "support@nftauction.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/collections/createNft": {
            "post": {
                "description": "保存 NFT 信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "保存 NFT 信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "nft名称",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "NFT描述",
                        "name": "description",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "NFT图片URL",
                        "name": "imageUrl",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "版税百分比",
                        "name": "royaltyPercentage",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "所属区块链网络id",
                        "name": "chainId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "类别id",
                        "name": "categorieId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "出价信息列表",
                        "schema": {}
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    }
                }
            }
        },
        "/collections/{address}/bids": {
            "get": {
                "description": "获取指定集合的所有出价信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "获取集合的出价信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "集合地址",
                        "name": "address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "过滤参数，JSON格式",
                        "name": "filters",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "出价信息列表",
                        "schema": {}
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    }
                }
            }
        },
        "/collections/{address}/history-sales": {
            "get": {
                "description": "获取指定集合的历史销售价格信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "获取历史销售价格信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "集合地址",
                        "name": "address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "链ID",
                        "name": "chain_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "7d",
                        "description": "时间范围(24h/7d/30d)",
                        "name": "duration",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "历史销售价格信息",
                        "schema": {}
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    }
                }
            }
        },
        "/collections/{address}/items": {
            "get": {
                "description": "根据过滤条件获取指定集合中的NFT物品列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "获取集合中的物品列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "集合地址",
                        "name": "address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "过滤参数，JSON格式",
                        "name": "filters",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "物品列表",
                        "schema": {}
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    }
                }
            }
        },
        "/collections/{address}/top-trait": {
            "get": {
                "description": "获取指定集合中物品特性的最高价格信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "获取物品特性的最高价格信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "集合地址",
                        "name": "address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "过滤参数，JSON格式",
                        "name": "filters",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "特性价格信息",
                        "schema": {}
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    }
                }
            }
        },
        "/collections/{address}/{token_id}": {
            "get": {
                "description": "获取指定物品的详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "获取物品详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "集合地址",
                        "name": "address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "物品Token ID",
                        "name": "token_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "链ID",
                        "name": "chain_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "物品详情",
                        "schema": {}
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    }
                }
            }
        },
        "/collections/{address}/{token_id}/bids": {
            "get": {
                "description": "获取集合中指定物品的所有出价信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "获取集合中特定物品的出价信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "集合地址",
                        "name": "address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "物品Token ID",
                        "name": "token_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "过滤参数，JSON格式",
                        "name": "filters",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "物品出价信息列表",
                        "schema": {}
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    }
                }
            }
        },
        "/collections/{address}/{token_id}/traits": {
            "get": {
                "description": "获取指定物品的特性(Attribute)信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "获取物品特性信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "集合地址",
                        "name": "address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "物品Token ID",
                        "name": "token_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "链ID",
                        "name": "chain_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "物品特性信息",
                        "schema": {}
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "errcode.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}